Offset Pruning for Pagination Queries (https://redirect.github.com/duckdb/duckdb/pull/19806)
This PR introduces basic row group pruning with using the query `LIMIT`
and `OFFSET` for pagination-style queries like `SELECT * FROM t ORDER BY
a LIMIT M OFFSET N`.

Currently, these queries are processed with a min-heap with size `M + N`
until a threshold of 0.7% of the table cardinality. For queries beyond
that threshold, the table is fully sorted and then limited. Especially
for larger tables, where we require out-of-core sorting, this results in
a hefty performance cliff:

Once again, we use the TSBS `cpu` table with 100M rows and a query
`SELECT * FROM cpu ORDER BY time LIMIT 100 OFFSET N`. To illustrate the
performance cliff, we set `N` once to 720000 and once to 730000 and
compare the nightly build to this PR.

| Offset | Nightly | This PR | Improvement |
| ------ | ------- | ------- | ----------- |
| 720K   | 0.09s   | 0.05s   | **1.8x**    |
| 730K   | 11s     | 0.05s   | **220x**    |
| 50M    | 11s     | 0.05s   | **220x**    |

The PR uses a similar algorithm to https://redirect.github.com/duckdb/duckdb/pull/19655 to exclude row groups. Note
that this is only possible, if there are no predicates and only for row
groups that do not contain null values in the sort column. Moreover,
this method only provides a benefit if the order column is already
sorted/partitioned to some degree, i.e., pruning is not possible if the
sort values are uniformly distributed.

Finally, this PR moves the pruning optimizations into its own optimizer
rule and adds the ability to fetch row group statistics from partition
stats in the optimization phase.

Upstream-commit: https://github.com/duckdb/duckdb/commit/aa588f86aec81c4b0d6ac5beb36a8cf31b7d7ced