[Compression] Prevent overriding `COMPRESSION_EMPTY` with `COMPRESSION_CONSTANT` (https://redirect.github.com/duckdb/duckdb/pull/19913)
This PR fixes https://github.com/duckdblabs/duckdb-internal/issues/6656

`COMPRESSION_EMPTY` was added by https://redirect.github.com/duckdb/duckdb/pull/15591 

`COMPRESSION_EMPTY` is used when we detect that the compression
algorithm used for the data can also encode the validity of the column.
So the validity does not need to be separately compressed.

`COMPRESSION_CONSTANT` is used when we detect that all values in the
segment are the same. This is also used for Validity in the case where
everything is null / everything is non-null.
So the data does not need to be compressed.

They are semantically similar, but some of the callbacks on the
CompressionFunction have different behaviors, for example `filter`.

This callback is used to implement filter pushdown at the storage level

For a constant-compressed validity segment this is either 0 or all for a
null/non-null filter.
For an empty-compressed validity segment this should be a no-op and
implemented by the column's compression function.

Upstream-commit: https://github.com/duckdb/duckdb/commit/f019064d5c92f4a594bd32c987df71de40e9ea63