[Storage] Implement storage for VARIANT columns (https://redirect.github.com/duckdb/duckdb/pull/19674)
Changes summary:

- Moved `VariantValue` from `extension/parquet/...` to core
- Moved `ShreddingState` to core, renamed as `VariantShreddingState`,
made it an abstract class
- Moved `VariantShredding` to core, made it an abstract class with pure
virtual method `WriteVariantValues`
- Moved the visitor struct created for `ConvertVariantToValue` to its
own header, renamed the struct to `ValueConverter`
- Moved the visitor struct created for `variant_normalize` to its own
header, renamed the struct to `VariantNormalizer`
- Created `VariantStats` for `BaseStatistics`
- Added `LogicalTypeId logical_type_id;` to `PersistentColumnData`[1]
- Created `VariantColumnData`

[1] This is needed to inform the de/serialization of `shredded` for
VariantColumnData

The internal structure of the `VariantColumnData` is:
```
STRUCT (
    unshredded VARIANT (as STRUCT),
    [ shredded <shredded type> ]
)
```

Where `shredded type` is of this shape:
```
STRUCT(
    untyped_value_index UINTEGER,
    typed_value <shredded type / shredded_type[] / primitive type>
)
```
This mimics the
[VariantShredding](https://github.com/apache/parquet-format/blob/master/VariantShredding.md)
of Parquet's VARIANT type, only we replace the repeated binary encoding
with an index (`untyped_value_index`) that refers to a value in the
`unshredded` field at the root.

At `Checkpoint` we determine what the shredded type of the Variant
should be, with an additional scan.
Then we scan the existing data, transforming it to the shredded
representation, write new `ColumnData` for it, which we then
`Checkpoint` (this is why `unshredded` is of type STRUCT, to not
recursively Checkpoint VARIANT data)

For debugging purposes the `force_variant_shredding` pragma was added,
which overrides the detected shredding type.

The `variant_minimum_shredding_size` pragma was also added, to set a
threshold to start shredding on Checkpoint, defaults to 30_000.
If the rowgroup contains less than `variant_minimum_shredding_size`
tuples, no shredding is applied to the Variant column.

Upstream-commit: https://github.com/duckdb/duckdb/commit/391b54ea59ef38e4cb5bcfbb37d059628f498180