Checkpoint: prevent adding blocks to the `free_list` that are still being used by concurrent readers (https://redirect.github.com/duckdb/duckdb/pull/19928)
Follow-up fix from https://github.com/duckdb/duckdb/pull/19867

This PR fixes the issue that was described in that PR:

> There are theoretically some more complications to this process -
namely if we were to do a CHECKPOINT twice old blocks could become
overwritten which would cause old RowGroupSegmentTrees to no longer be
valid, as they would point to blocks that have been overwritten with new
data. However, it is not possible to hit this issue currently because we
still have very aggressive locking around writes and checkpoints - it's
not actually possible to start a second CHECKPOINT while a read of old
data is still active in the current setup. However, this is an issue
that will need to be dealt with as we relax the write locks around
checkpoints in future PRs.

I did not believe that this could occur yet given the current aggressive
locking around checkpointing, but it could sporadically occur when doing
optimistic writes. This could occur sporadically in the
`test/sql/storage/checkpoint/concurrent_load_delete.test_slow` test -
causing it to hit an assertion.

This PR fixes the issue by checking after the checkpoint which blocks
are still alive (i.e. which blocks have not yet been unregistered using
`BlockManager::UnregisterBlock` - which is called in the destructor of
the `BlockHandle`). For any blocks that are still in-use, these are not
directly added to the in-memory free list but rather added to the
`free_blocks_in_use` list.

These blocks are still written as free blocks to the file - meaning on a
restart they are correctly labeled as free blocks again. However, while
there are still scans that might be reading these blocks, these ids will
not be handed out by `GetFreeBlockId` - meaning they will not be
overwritten while threads might potentially be reading from them. Upon
being fully unregistered, these blocks are then added to the free list
so that they can be re-used again.

This also affects truncating of the file - both for truncating free
blocks at the end of the file, and truncating blocks in the middle of
the file. We will only truncate the file during a checkpoint if blocks
at the end of the file are fully free and not referenced by any active
threads either.

This PR also has a minor fix for `CatalogSet::AlterEntry`. Effectively
when doing a WAL replay of a database with foreign key constraints we
would replay the foreign key constraint adding without a client context.
This would not be cleaned up correctly, causing the old copy of the
table to be kept alive. This would then conflict with the new block
handle tracking and prevent blocks from being cleaned up correctly.

Upstream-commit: https://github.com/duckdb/duckdb/commit/0e4723ab81e5c737bada86672b2761e52e2a0109