Avoid eagerly grabbing the checkpoint lock when performing inserts, instead grab a shared table lock (https://redirect.github.com/duckdb/duckdb/pull/19937)
Currently whenever a database is intended to be modified using a
statement such as `INSERT`, `DELETE`, etc we eagerly grab the
`SharedCheckpointLock` and store in the `DuckTransaction`. As long as
that transaction lives, we can then not initiate a checkpoint. This can
keep on applying also after the transaction has concluded as the lock is
only released when the transaction is cleaned up, which only happens
when there are no active transactions that depend on the data stored
within that transaction anymore. If there are readers that need an old
snapshot of the data for a long time, these transactions might not be
cleaned up for a long time as well.

This has two negative effects in concurrent scenarios with several
writers, or even a single writer and a single (separate) checkpointer:
* Checkpoints can fail to be performed, as the checkpoint lock cannot be
grabbed. As a result, we might not checkpoint for extended periods of
time, leading to the WAL file continuously growing, data not being
compressed / vacuumed, etc.
* When a checkpoint is performed (either forcibly through `FORCE
CHECKPOINT`, or when the lock manages to be grabbed) writers must wait
for the checkpoint to complete, preventing them from continuing
operations and potentially introducing large latency spikes.

This PR makes an initial move towards grabbing the checkpoint lock less
and allowing more operations to happen while a checkpoint is performed
(and vice versa - allowing a checkpoint to start in more scenarios).
This PR removes the eager grabbing of the `SharedCheckpointLock` for
appends (through e.g. an `INSERT INTO` statement). This is accomplished
by passing the intended database modification to `RegisterDBModify`.

If the modification is an append, we allow it to proceed without
grabbing the shared checkpoint lock. Instead, when performing an append
on a table, we grab a shared lock to the table (that was previously used
for scans). Checkpoints can then still start, but need to wait for an
append to finish on a table level (and vice versa).

Note that this is still rather limited - because we still grab the
`SharedCheckpointLock` when writing to the WAL. That means that we still
cannot *commit* transactions that have inserted data while a checkpoint
is happening (and cannot start a checkpoint while a transaction that
inserted is writing to the WAL). I plan to tackle that in a follow-up
PR.

Upstream-commit: https://github.com/duckdb/duckdb/commit/f38e0ff7c3870924ff5f4a5c01ee313b45558f37