[Checkpoint] Fully rewrite stats on Checkpoint (https://redirect.github.com/duckdb/duckdb/pull/19932)
This PR fixes https://github.com/duckdblabs/duckdb-internal/issues/6641

Previously the new stats that are created during Checkpoint were
initialized with a copy from the existing stats.
These existing stats could contain data that has since been deleted, so
this way the stats of the table could only *grow*, i.e min and max could
never become higher and lower respectively than any data that was ever
in the table.

Next to that problem, the Checkpoint was only serializing the new stats
to disk, but not immediately updating the stats of the
RowGroupCollection that was checkpointed.

In the tests this resulted in requiring a `restart` to make sure the new
stats were visible, but as a user this is likely more impactful because
restarting is likely not commonplace or in an in-memory situation that's
not even possible.

Upstream-commit: https://github.com/duckdb/duckdb/commit/f8b2dc561bcaa82868f8791d637fc466abe2be0f