Backport revert append fixes (https://redirect.github.com/duckdb/duckdb/pull/19941)
In https://github.com/duckdb/duckdb/pull/19937 I fixed an issue with
`RevertAppend` potentially leaving the segment tree in an undefined
state in certain situations (when we would truncate at exactly the row
group boundary). This method is called when a unique / primary key
constraint violation is found during the commit process. This could then
cause issues to happen in subsequent code. This PR backports the fix for
this to v1.4.

Upstream-commit: https://github.com/duckdb/duckdb/commit/c3fb0e6beb93158cc765d6027da6cf023efd1c40