[Stats] Improve the `VariantStats` class to support merging and better visualization (https://redirect.github.com/duckdb/duckdb/pull/19950)
This PR is a continuation on https://redirect.github.com/duckdb/duckdb/pull/19674 

The `VariantStats` created there was too simple, as it represented the
shredding state as a single boolean:
```c++
bool is_shredded;
```

What previously happened was that `is_shredded` would be initialized to
false.
When the shredded stats were added, or on explicit creation of the
shredded stats, this turns `is_shredded` to true.

### Merge problem

In the `Merge` of the stats, we would accept merging `is_shredded ==
false` with `is_shredded == true`, basically acting like an upgrade,
turning `is_shredded` to true.

Then `is_shredded == true` could be merged with another `is_shredded ==
true`, and if those are incompatible, `is_shredded` reverts back to
false.
If we then merge another `is_shredded == true` into it, it becomes
shredded again.

That logically makes no sense.
Shredded stats only make sense if we can be certain that it contains
*all* information about the variant type we're shredding on.

### Solution

`is_shredded` has been replaced with `VariantStatsShreddingState`:
```c++
enum class VariantStatsShreddingState : uint8_t {
	//! Uninitialized, not unshredded/shredded
	UNINITIALIZED,
	//! No shredding applied yet
	NOT_SHREDDED,
	//! Shredded consistently
	SHREDDED,
	//! Result from combining incompatible shreddings
	INCONSISTENT
};
```

The matrix of merges is as follows:

| Type | UNINITIALIZED | NOT_SHREDDED | SHREDDED | INCONSISTENT |

|----------------|----------------|---------------|---------------|--------------|
| UNINITIALIZED | UNINITIALIZED | NOT_SHREDDED | SHREDDED | INCONSISTENT
|
| NOT_SHREDDED | INCONSISTENT | INCONSISTENT | INCONSISTENT |
INCONSISTENT |
| SHREDDED | INCONSISTENT | INCONSISTENT | SHREDDED/INCONSISTENT[1] |
INCONSISTENT |
| INCONSISTENT | INCONSISTENT | INCONSISTENT | INCONSISTENT |
INCONSISTENT |

[1] `SHREDDED` + `SHREDDED` can only stay `SHREDDED` if the shredding
types are compatible.

VariantStats starts off as being `UNINITIALIZED`, can be manually marked
as `NOT_SHREDDED` (as happens in Append) or upgraded to `SHREDDED` when
uninitialized/not_shredded and shredded information is added (this can
only happen before any merges occur)

### `variant_extract` stats propagation

We've also implemented the propagation of the variant stats through the
`variant_extract` method, so child stats of shredded fields are
correctly propagated, to be used later by projection pushdown
optimizations.

### VariantStats ToString()

The previous ToString of VariantStats was a bit of an information dump,
which was hiding valuable data, but had to be pieced together manually.
We've removed that, in favor of just displaying whether the field/column
is fully shredded, as this is the only information you really care about
with VariantStats anyways.

Upstream-commit: https://github.com/duckdb/duckdb/commit/06f5a6dd6923a417937a6536a5f9626e881925db