[Parquet] Fix prepared copy option parameter (https://redirect.github.com/duckdb/duckdb/pull/19966)
This PR fixes https://github.com/duckdb/duckdb/issues/19884

I've added a check to see if the expression is foldable, otherwise it
can't be used in EvaluateScalar.
If it's not foldable, we'll check if it has a parameter expression and
throw ParameterNotResolved if that's the case, otherwise we throw a
BinderException.

Upstream-commit: https://github.com/duckdb/duckdb/commit/1cdb7e77a28cca9f8ab953103f0ba4ea43461a0c