Unbound index binding with context (https://redirect.github.com/duckdb/duckdb/pull/19953)
If we have a restart -> checkpoint -> restart, it is possible that there
are buffered unbound index operations that have not been replayed, and
are currently being dropped. During checkpointing we may or may not have
a context. This PR fixes the scenario where there is a context + active
transaction, as we can just bind the indexes before serialization to
replay the buffered operations. (A separate PR is in the works for if
there isn't a context, which requires serialization of unbound indexes).

Technically the unbound index serialization PR would fix the context
case as well, but this is going in as a separate PR since it is better
to just bind the indexes and replay buffered operations if we have the
context available, instead of serializing them to disk.

Upstream-commit: https://github.com/duckdb/duckdb/commit/136dd6ada5619cd9293a03cca717515f7436cfa8