CLI: Fully unify result rendering, and fix automatic pager mode (https://redirect.github.com/duckdb/duckdb/pull/19988)
This PR fully unifies the result rendering code, and uses the newly
unified result rendering to fix up the automatic pager mode. Instead of
guessing whether or not a result is wide by the amount of columns, we
now actually measure how wide the result set will be and enable the
pager based on that.

We also do some random clean-up, e.g. removing `QueryResult::
MoreRowsThan` again and moving the iteration / caching to the shell
renderers.

Upstream-commit: https://github.com/duckdb/duckdb/commit/cbb385c18c1fad9b1b07f9789e674593119b17fd