CLI: Various auto-complete fixes (https://redirect.github.com/duckdb/duckdb/pull/19995)
* Fix a rendering bug with auto-complete that would cause it to
continuously clear upwards if the buffer did not start at the top of the
terminal
* Differentiate between `fuzzy suggestions` and `exact suggestions`,
where exact suggestions have an exact prefix match with the text, and
fuzzy suggestions do not. Then show a smaller number of max fuzzy
suggestions. This prevents polluting the auto-complete output with
(likely) unrelated files.
* Make Ctrl+C cancel auto-complete, instead of clearing the entire
buffer

Upstream-commit: https://github.com/duckdb/duckdb/commit/a692afcd3c294adbe294e50f7f8cf59b24aaefca