[Art][Wal]Unbound index allocations (https://redirect.github.com/duckdb/duckdb/pull/19901)
Follow up to https://github.com/duckdb/duckdb/pull/19477, fix for
https://github.com/duckdblabs/duckdb-internal/issues/6613

The previous PR added support for buffering and replaying WAL index
deletes, however that introduced a memory over-allocation issue, as the
UnboundIndex was storing a vector of BufferedIndexData, which stored
each buffered operation in a ColumnDataCollection. This was extremely
wasteful because if there are interleavings (insert -> delete -> ...) a
single operation to be replayed would be stored in a
ColumnDataCollection with an internal allocation of
STANDARD_VECTOR_SIZE.

EDIT: See @Mytherin's comment below, this PR fixes the issue by changing
the way buffering works, now we use two buffers, one for inserts, and
another for deletes. Since the inserts and deletes may be interleaved,
however, we need an additional vector<ReplayRange> data structure that
stores replay operations and their intervals within the respective
buffer. This all stored in BufferedIndexReplays within UnboundIndex.

Buffering data is much simpler now, as we can just append directly to
either the insert or delete ColumnDataCollection, as well as appending a
ReplayRange node (or extending the range of the last node, if the replay
operation is the same type of operation).

Replaying is more efficient now, as we now maintain two interleaved
scans on the respective contiguous ColumnDataCollections, fetching one
DataChunk at a time to replay.

Upstream-commit: https://github.com/duckdb/duckdb/commit/52fe0d2bffdc766e7a75a9f966c6db537e3ffdca