[Compression] Fix an issue with NULL updates to a column compressed with DICT_FSST (https://redirect.github.com/duckdb/duckdb/pull/20009)
This PR fixes https://github.com/duckdblabs/duckdb-internal/issues/6656

The problem was that during CHECKPOINT, we check if the ColumnData has
any changes, which for DICT_FSST would mean the question: "are the
strings changed?"
If that answer is no, but the validity *has* changed, we should still
rewrite the data.

This mimics behavior on `main`, which also rewrites everything when
either validity or data has changes.

Upstream-commit: https://github.com/duckdb/duckdb/commit/4b056d722eb9bb950aeb51afda4bc70cfcd8b8b8