Add a secondary extension directory for package managers (https://redirect.github.com/duckdb/duckdb/pull/17499)
This add the possibility to store duckdb extensions in a secondary
directory. The main intention is to allow package managers like conda,
pixi or spack to install DuckDB extensions. This is especially relevant
in cases where you run in air-gapped environments where you might have a
mirror of these package repositories but are not able to connect to
DuckDB's extension registry.

See also https://github.com/duckdb/duckdb/discussions/15352 for the
initial discussion.

Upstream-commit: https://github.com/duckdb/duckdb/commit/3c4456b3e3bc308ad6b173ddf5a991903e16bfab