Enable windows_arm64 arch for main extensions (https://redirect.github.com/duckdb/duckdb/pull/20004)
This change adds `opt_in_archs` argument to the `Extensions.yml`
workflow and passes it down to `_extension_distribution.yml`.

The default value is set to `windows_arm64;` with the intention to keep
it there (if it will work correctly and will not cause problems) to have
Windows ARM64 extension builds enabled by default.

This change only enables it for Main extensions. Rust-based extensions
(`delta`) can be added later - it requires Python wheel re-build (for
`configure` step, until `windows_arm64` binary wheel is available). It
works on CI, but takes additional 35 mins.

Upstream-commit: https://github.com/duckdb/duckdb/commit/bea03d5c8bfb130994d97852b185f84520cfb6db