propagate gate status in Node4::DeleteChild (https://redirect.github.com/duckdb/duckdb/pull/20044)
Bugfix for https://github.com/duckdb/duckdb/issues/19975

Propagate gate status in Node4::DeleteChild, since Prefix::Concat logic
depends on both the prev_node4 gate status (was the node4 being freed a
gate node), as well as the overall gate status (did we enter through a
gate at some point earlier). The latter was not being propagated
correctly, and node compression logic for nested ARTs storing rowids
depends on this, since we do not need to maintain prefix chains, as they
are implicit in the rowid's.

Unfortunately it is not possible to create a minimal test for this
scenario, however once
https://github.com/duckdblabs/duckdb-internal/issues/6083 is solved,
there is a large reproducer for this scenario that I downloaded from the
person who opened this issue (and the data/sql file is not confidential,
so can be used).

Thank you @taniabogatsch for dealing the final blow to this bug! 

Upstream-commit: https://github.com/duckdb/duckdb/commit/7f0b18a746b7df524c26ed5b21dd1bad441dbac1