CLI: Improve rendering of wide values, and improve auto-complete (https://redirect.github.com/duckdb/duckdb/pull/20074)
* When shortening wide columns, prefer to shorten the widest columns
first before shortening less wide columns, instead of shortening
everything by an equal percentage. This means that if we have a table
that is e.g. `name, description` - where `name` is slightly above the
"wide" threshold (20 bytes) and "description" is significantly above
this threshold, we will only shorten the description field - keeping
`name` readable.
* When stretching out values across rows, give all rows the same amount
of space for stretching out, instead of greedily stretching out the
first rows
* Don't stretch out columns across rows if there are hidden columns in
the result set
* Always add a separator between rows if we are stretching out columns
* For auto-complete, use only the match score instead of the "adjusted
sort score" when determining if there are ties or not. This way if we
have a directory that has e.g. `my_dir/` and `my_file.parquet` as
auto-complete options, `my_` will not instantly auto-complete to
`my_dir/` but instead consider this a tie. `my_dir/` will still appear
first in the sorted suggestion list, however.

Upstream-commit: https://github.com/duckdb/duckdb/commit/921363ad08e424a114c030bd6eae23c61881a97e