Checkpoint: Allow vacuuming of rows for tables that have indexes if the vacuum does not change the rowids (https://redirect.github.com/duckdb/duckdb/pull/20073)
Currently vacuuming is disabled for indexes entirely because vacuum has
the potential of changing row-ids. This prevents deleted rows from being
cleaned up, always leaving deletion markers in place. This is necessary
because the index relies on row-ids, and if we change the row-id of a
row the index no longer points to the correct row.

The current behavior with vacuuming is unnecessarily restrictive -
however. In certain cases, vacuuming does not change row-ids of existing
rows. For example:

* We can merge adjacent row groups if they do not have deletes, e.g. if
we have two row groups:
```
[RowGroup 1]
    start: 0
    count: 10
[RowGroup 2]
    start: 10
    count: 10
```

We can merge them into one row group:

```
[RowGroup 1]
    start: 0
    count: 20
```

This does not change the row ids of any rows, so it is valid also if we
have indexes.

* We can remove row groups at the tail of a table. For example:

```
[RowGroup 1]
    start: 0
    count: 10
[RowGroup 2]
    start: 10
    count: 10
    deleted: 10
```

If we remove `RowGroup 2`, no existing rows have their row-ids changed,
so we can just delete it. This solves the issue that truncating a table
with an index on it does not delete any of the actual rows.


This doesn't fully solve the vacuuming issue with indexes present but
solves it in some common scenarios.

Upstream-commit: https://github.com/duckdb/duckdb/commit/104c8d0842aa31e2d1e31bb2d29fe5ae4c8c4b52