Remove undefined loop (https://redirect.github.com/duckdb/duckdb/pull/20105)
This loop is undefined behavior since info and parent.info are both
shared_ptr's to the same DataTableInfo, and thus pointing to the same
vector in the loop iteration, which can possibly cause a reallocation
with the push_back. The standard only says that references before an
insert that causes a reallocation are valid, so anything goes here.

More generally, however, IndexStorageInfo's are not used at this point,
as they should be already contained in UnboundIndexes (which are
immediately bound at the start of this constructor), or if the indexes
were already bound, they can stop being used there (and then later on
they are recreated again from bound indexes for serialization).

Will most likely make a separate PR on main that more explicitly manages
the lifetimes of IndexStorageInfos to make this clear.

Upstream-commit: https://github.com/duckdb/duckdb/commit/01198fffc2b27c0d197f6ac2879d5da7f2a75ff7