Fix concurrent test failure (https://redirect.github.com/duckdb/duckdb/pull/20129)
This test is sporadically failing in CI due to the test itself being
wrong.

We write, in-order, to `main_db.integers` followed by writing to
`other_db.ranges`. Depending on when the abort happens, data can be
either (1) in `main_db.integers`, (2) in both `main_db.integers` and
`other_db.ranges`, or (3) in neither of them. It is therefore completely
valid for the data not to be in `other_db.ranges`. The current code
assumes at least *some* data makes it to `other_db.ranges` and fails if
that table is empty - but that is not always the case. Depending on when
the abort is triggered that table might be empty. This PR fixes the test
to reflect that correctly.

Upstream-commit: https://github.com/duckdb/duckdb/commit/2a8be25a4e988217d728bc792b0ad8c49ab3a79c