Headers missing from http logs (https://redirect.github.com/duckdb/duckdb/pull/20030)
The extra_headers (settable through HTTP type secrets) would not show up
in the http logs.

After this PR, instead of merging the extra_headers into the requests on
every request, we just copy all headers into the requestinfo. This pays
a small price of copying all headers for each request, but `select
request.headers from duckdb_logs_parsed('HTTP');` will now correctly
show the headers that are added to the request. I think this is worth
the cost given the context of a network request and how confusing it is
to have requests be sent with invisible headers.

### possible optimization
We can improve this by only copying the headers when extra_headers are
present and having the requestinfo hold either a owning pointer or a
non-owning pointer

Upstream-commit: https://github.com/duckdb/duckdb/commit/1c03ec08121a5a3b9aa2e10d9de17dbb144042b5