Provide option disable cache validation (https://redirect.github.com/duckdb/duckdb/pull/19886)
This is a feature request I've been longing for a while.
It's common to use duckdb for query on immutable parquet files and use
external file cache to store in-memory (without extensions); when I was
still with mooncake, we used to suffer thousands of HEAD requests for
immutable remote parquets, which hurt latency and ran out of connection.

This PR exposes `validate_external_file_cache` file option (which is by
default set for
[ducklake](https://github.com/duckdb/ducklake/blob/51bfd079001964d431d592bbac2a351ad785e52a/src/storage/ducklake_multi_file_list.cpp#L351))
as a duckdb config, so users (who know the workload pattern best) could
opt out to skip necessary validation.

Upstream-commit: https://github.com/duckdb/duckdb/commit/9743dd21e18921f1b325af5c74c28ad4f0d1bf12