Allow concurrent insertions into tables with indexes while checkpointing (https://redirect.github.com/duckdb/duckdb/pull/20160)
Follow-up from https://github.com/duckdb/duckdb/pull/20052 and
https://github.com/duckdb/duckdb/pull/20119

When doing a checkpoint, we need to write the state of the index **as of
the checkpoint transaction id** to disk. Currently, this is achieved by
preventing concurrent modifications to the index while a checkpoint is
running. While this works, it is limiting. We want to allow concurrent
modifications to the table also while a checkpoint is running. This is
already possible for tables without indexes since the previous PR. This
PR removes the extra restriction placed on tables with indexes.

The way this works is by extending the `IndexEntry` with two new
structures:

```cpp
struct IndexEntry {
    // ...
	unique_ptr<BoundIndex> added_data_during_checkpoint;
	optional_idx last_written_checkpoint;
};
```

**Appends.** When appending to an index, we check if an active
checkpoint is running. If an active checkpoint is running, and
`last_written_checkpoint` is not set to the active checkpoint id, we
append to the delta index `added_data_during_checkpoint` instead. This
prevents concurrent modifications from modifying the main index. For
unique / primary key constraint indexes, we still need to do a look-up
in the main index prior to inserting into the delta index, however.

```cpp
optional_ptr<BoundIndex> append_index = main_index;
optional_ptr<BoundIndex> lookup_index;

if (active_checkpoint.IsValid()) {
	// active checkpoint
	if (!entry.last_written_checkpoint.IsValid() ||
	    entry.last_written_checkpoint.GetIndex() != active_checkpoint.GetIndex()) {
	    // need to append to the delta index
		if (!entry.added_data_during_checkpoint) {
			entry.added_data_during_checkpoint = CreateNew();
		}
		if (main_index.IsUnique()) {
			// need to look-up prior to appending
			lookup_index = main_index;
		}
		append_index = entry.added_data_during_checkpoint;
	}
}
```

**Lookups.** When doing an ART look-up, we need to look in
`added_data_during_checkpoint` as well if it is present. The same
applies when doing verification of foreign key constraints.

**Checkpoints.** When doing a checkpoint, we write the index to disk.
After the checkpoint is finished - we look over all indexes, and merge
any delta indexes into their corresponding main index and set
`last_written_checkpoint` to the current checkpoint id. After this
point, new data will be written directly to the main index again - at
least until the next checkpoint is started.

For now, deletions do not need to take `added_data_during_checkpoint`
into account - as deletions still work with the old "stop the world"
approach - i.e. a deletion cannot occur while a checkpoint is running at
this moment. We plan to address this in a future PR.



### LocalTableStorage::AppendToIndexes Split-Up

Previously, this method would interleave appending to indexes and
appending to the table, to avoid having to do two scans of the
`RowGroupCollection`. This made the code pretty complex. In particular
when a constraint violation was found, as we would need to revert both
appends made to indexes and to the table. The method was also very
confusingly named given that it would append to **both** indexes **and**
to the base table.

This PR splits this method into `LocalTableStorage::AppendToIndexes` and
`LocalTableStorage::AppendToTable`. We now only start appending to the
table after all indexes have been appended to - which prevents us from
having to revert the table appends if a constraint violation was found.
This greatly simplifies the code. It is also not very clear if this cost
any performance - as (1) scanning a `RowGroupCollection` is not
particularly expensive compared to all of the other operations we are
doing (inserting into indexes, etc) and (2) this method is only used for
smaller inserts anyway, as larger inserts use the optimistic writing
code path.


### `RowGroupCollection::RevertAppendInternal` rework

If an error occurs during commit, we need to revert any changes made to
the table in order to safely turn the commit into a rollback. For
appends, this happens through
`RowGroupCollection::RevertAppendInternal`. This method would previously
directly remove data that was appended. While this generally works fine
- it can cause issues when other queries are concurrently scanning the
table. Doing a destructive in-place modification of a `SegmentTree` is
not thread-safe. While no other transaction would need to scan the row
groups that were appended (as the append never succeeded) - they could
have been looking at the row group to determine if they need to scan it
or not. This can cause a data race issue.

This PR reworks that method to instead construct a new
`RowGroupSegmentTree` that only includes the relevant segments, and then
safely overrides the set of row groups with the new (pruned) set of row
groups, solving this issue.

Upstream-commit: https://github.com/duckdb/duckdb/commit/efe80f478d5a619dd588434717fc3d4eb66668f9